+++ DAU-API/GameTest/utils.h
@@ -10,9 +10,9 @@ enum
     ANIMATE
 };
 
-std::vector<CSimpleSprite *> generateMeteors(float level);
+std::vector<CSimpleSprite *> generateMeteors(int level);
 void gameOver(CSimpleSprite *player, std::vector<std::vector<CSimpleSprite *>> &allMeteors, bool &lose);
 bool outOfBounds(float x, float y);
-void updateMeteors(std::vector<std::vector<CSimpleSprite *>> &allMeteors, float level);
+void updateMeteors(std::vector<std::vector<CSimpleSprite *>> &allMeteors, int level);
 
 #endif
